created a funnel chart
for creating it.. created a new calculated filed with value of -(the value your mapping). Then placing it in the right way and changing automatic to area. Removing axis from down and then selecting all in the marks section and clicking on tool tip and removing funnel(new calculated field) from it so that there's no counter intuative clutter in the sheet.