Created 'Sales trend with forecast' sheet
created a sheet with order date and sales. Included a forecast in it. To include forecast go to Analytics tab and drag and drop the forecast icon to the white sheet where you get the option to add a forcast. After that you can include trend lines by right clicking on empty sheet and clicking on show trend lines.
NOTE :- For forecast you need atleast 12 months of data :+1: